added WebSockets to the flask app - next plan is to build board front end with react that can talk to websockets in the flask server. then will connect flask server to the c++ game server using berkeley style sockets
