+++ game-server/client/client_api.hh
