+++ game-server/client/view.cc
