+++ game-server/client/view.hh
