+++ game-server/server/exitpipe.cc
