+++ game-server/server/exitpipe.hh
