+++ game-server/shared/utilities.cc
