+++ game-server/shared/utilities.hh
