--- landing-page/docker-compose.yml
@@ -1,13 +1,0 @@
-services:
-  api:
-    build:
-      context: .
-      dockerfile: Dockerfile.api
-    image: react-flask-app-api
-  client:
-    build:
-      context: .
-      dockerfile: Dockerfile.client
-    image: react-flask-app-client
-    ports:
-      - "3000:80"
