+++ web-app/api/requirements.txt
@@ -1,5 +1,6 @@
 click==8.0.1
 Flask==2.0.1
+flask-sock==0.4.0
 gunicorn==20.1.0
 itsdangerous==2.0.1
 Jinja2==3.0.1
