+++ web-app/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "landing-page",
+  "name": "web-app",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
