+++ web-app/public/index.html
@@ -28,6 +28,33 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
+    <h1>Flask-Sock Demo</h1>
+    <div id="log"></div>
+    <br>
+    
+    <form id="form">
+      <label for="text">Input: </label>
+      <input type="text" id="text" autofocus>
+    </form>
+
+    <script>
+      const log = (text, color) => {
+        document.getElementById('log').innerHTML += `<span style="color: ${color}">${text}</span><br>`;
+      };
+
+      const socket = new WebSocket('ws://localhost:5000/echo');
+      socket.addEventListener('message', ev => {
+        log('<<< ' + ev.data, 'blue');
+      });
+      document.getElementById('form').onsubmit = ev => {
+        ev.preventDefault();
+        const textField = document.getElementById('text');
+        log('>>> ' + textField.value, 'red');
+        socket.send(textField.value);
+        textField.value = '';
+      };
+    </script>
+
     <div id="root"></div>
     <!--
       This HTML file is a template.
@@ -41,3 +68,10 @@
     -->
   </body>
 </html>
+
+<head>
+  <title>Flask-Sock Demo</title>
+</head>
+<body>
+  
+</body>
