+++ web-app/src/App.js
@@ -1,5 +1,4 @@
 import React, {useState, useEffect} from 'react';
-import logo from './logo.svg';
 import './App.css';
 
 function App() {
@@ -14,17 +13,13 @@ function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
         <a
           className="App-link"
-          href="https://reactjs.org"
+          href="http://localhost:3000/api/game"
           target="_blank"
           rel="noopener noreferrer"
         >
-          Learn React
+          Join Game
         </a>
         <p>The current time is {currentTime}.</p>
       </header>
