+++ web-app/src/game-page/Game.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function Game() {
+
+  return (
+    <div className="Game">
+        <p>The current time is {currentTime}.</p>
+    </div>
+  );
+}
+
+export default Game;
