+++ web-app/src/game-page/game-index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './game-index.css';
+import Game from './Game';
+
+ReactDOM.render(
+  <React.StrictMode>
+    <Game />
+  </React.StrictMode>,
+  document.getElementById('root')
+);
+
